% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas}
\alias{canvas}
\title{a title}
\usage{
canvas(obj, opts = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{obj}{data.frame that contains coordinates x,y}

\item{opts}{opttions to pass to widget, see details}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}
}
\description{
a description.
}
\details{
options to pass to server side includes:animation, animation duration, animation ease, animation radius,
  duration which is the speed which the animation moves (default 10000ms). 
When deploying a shiny app the number of samples taken from the path is a function of the size of the duration.
and loop if set to TRUE (default FALSE) automatically restarts the animation when it reaches the end.
pathRadius set the size of the circle that moves along the interpolated curve.
}
\examples{
if(interactive()) canvas(obj=data.frame(x=1:10,y=runif(10)))

}
