% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidSpline.R
\name{fluidSpline}
\alias{fluidSpline}
\title{<Add Title>}
\usage{
fluidSpline(n = 11, interpolation = "linear", cW = 1000, cH = 500,
  width = NULL, height = NULL, elementId = NULL)
}
\description{
<Add Description>
}

